pyproject.toml
[tool.poetry]
name = "love-vids-backend"
version = "0.1.0"
description = "VibeVids.ai - AI-powered video gift platform backend"
authors = ["Antigravity <ai@vibevids.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.26.0"
stripe = "^7.13.0"
razorpay = "^1.4.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
boto3 = "^1.34.23"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
